* [SDRBaseband, rxs-parsing-core, USRPFrontEnd] Pilot CSI interation done

+++ CSISegment.hxx
@@ -115,7 +115,7 @@ public:
     uint8_t numTx = 1;
     uint8_t numRx = 1;
     uint8_t numESS = 0;
-    uint8_t numCSI = 1;
+    uint16_t numCSI = 1;
 };
 
 class CSI {
